+++ src/daft-warc/src/lib.rs
@@ -613,3 +613,68 @@ fn combine_stream<T, E>(
         }
     })
 }
+
+#[cfg(test)]
+mod tests {
+    use std::sync::Arc;
+
+    use common_error::DaftResult;
+    use daft_core::prelude::{Field, Schema, TimeUnit};
+    use daft_io::{IOConfig, IOStatsContext};
+
+    use crate::{read_warc_bulk, WarcConvertOptions};
+
+    #[test]
+    fn test_warc_read_iostats() -> DaftResult<()> {
+        let schema = Arc::new(Schema::new(vec![
+            Field::new("WARC-Record-ID", daft_core::prelude::DataType::Utf8),
+            Field::new("WARC-Type", daft_core::prelude::DataType::Utf8),
+            Field::new(
+                "WARC-Date",
+                daft_core::prelude::DataType::Timestamp(
+                    TimeUnit::Nanoseconds,
+                    Some("Etc/UTC".to_string()),
+                ),
+            ),
+            Field::new("Content-Length", daft_core::prelude::DataType::Int64),
+            Field::new(
+                "WARC-Identified-Payload-Type",
+                daft_core::prelude::DataType::Utf8,
+            ),
+            Field::new("warc_content", daft_core::prelude::DataType::Binary),
+            Field::new("warc_headers", daft_core::prelude::DataType::Utf8),
+        ])?);
+        let io_config = Arc::new(IOConfig::default());
+        let io_client = daft_io::get_io_client(true, io_config)?;
+        let io_stats = IOStatsContext::new("test_warc_read");
+        let warc_file = format!("{}/test/example.warc", env!("CARGO_MANIFEST_DIR"),);
+        let warc_gz_file = format!("{}/test/example.warc.gz", env!("CARGO_MANIFEST_DIR"),);
+
+        let convert_options = WarcConvertOptions {
+            schema: schema.clone(),
+            predicate: None,
+            include_columns: None,
+            limit: None,
+        };
+
+        let _ = read_warc_bulk(
+            &[&warc_file, &warc_gz_file],
+            convert_options,
+            io_client,
+            Some(io_stats.clone()),
+            true,
+            None,
+            8,
+        )?;
+
+        let warc_file_size = std::fs::metadata(&warc_file)?.len() as usize;
+        let warc_gz_file_size = std::fs::metadata(&warc_gz_file)?.len() as usize;
+        let bytes_read = io_stats.load_bytes_read();
+        assert_eq!(
+            bytes_read,
+            warc_file_size + warc_gz_file_size,
+            "IO stats should record the bytes read correctly"
+        );
+        Ok(())
+    }
+}
